Moved apollo config to api folder
