+++ src/index.tsx
@@ -4,42 +4,10 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import 'material-symbols';
-import { ApolloClient, InMemoryCache, ApolloProvider, gql} from '@apollo/client';
+import { ApolloProvider} from '@apollo/client';
+import { client } from './api/client';
 import { BrowserRouter } from 'react-router-dom';
-import { GraphQLWsLink } from '@apollo/client/link/subscriptions';
-import { createClient } from 'graphql-ws';
-import { split, HttpLink } from '@apollo/client';
-import { getMainDefinition } from '@apollo/client/utilities';
 
-
-const httpLink = new HttpLink({
-  uri: 'http://localhost:8080/graphql',
-  headers: {
-    Authorization: `Basic ${btoa('admin:password')}` // Adds a header for authorization
-  }
-});
-export const wsLink = new GraphQLWsLink(createClient({
-  url: 'ws://localhost:8080/graphql',
-}
-));
-
-const splitLink = split(
-  ({ query }) => {
-    const definition = getMainDefinition(query);
-    return (
-      definition.kind === 'OperationDefinition' &&
-      definition.operation === 'subscription'
-    );
-  },
-  wsLink,
-  httpLink,
-);
-
-
-export const client = new ApolloClient({
-  link: splitLink,
-  cache: new InMemoryCache(),
-});
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
